Tweaked the regular expression for detecting USB connects.
The general pattern for a USB device name seems to be N-(N.?)+ where N is a number and period is a period.

For instance we can have just "3-1", but also (when using a hub, I think) "1-4.1" or "1-5.2.3".

This needs to happen throughout, and probably some code needs to wisen up, too.